+++ test/Parity Tests/Aggregate Tests/Aggregate Tests.js
@@ -35,6 +35,7 @@ module.exports = function ( Driver )
 		require( './test-suite/Date Operator Tests.js' )( Driver );
 		require( './test-suite/Set Operator Tests.js' )( Driver );
 		require( './test-suite/Array Operator Tests.js' )( Driver );
+		require( './test-suite/Object Operator Tests.js' )( Driver );
 		require( './test-suite/Stage and Accumulator Tests.js' )( Driver );
 		require( './test-suite/Expression Rejection Tests.js' )( Driver );
 	} );